ARM: uncompress.h: Remove unused arch_decomp_wdog defines
arch_decomp_wdog is only used by lib/inflate.c which isn't used
by the ARM arch since 2.6.33.
Stop carrying it around and re-adding it with every new machine.

The only real user was the Samsung platform, but since this
feature wasn't working for over a year, it's probably
not needed at all.

Signed-off-by: Zoltan Devai <zoss@devai.org>